Added contouring and some tracking to video playback
It will now draw a contour around a detected object. If the object has a
combined area of over 500 pixels, it will note this and draw a red dot
in the middle of a green rectangle that forms around the object. The
goal is for this red dot to check to see if it collides with a bar and
in what direction it travels to see if a car is entering or leaving a
parking lot.